+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.2.0] - 2021-04-15 - Poor Unfortunate Types 🦑
+
 ### Fixed
 
 - Converting a ternary assigned to a variable used to remove other eventual declarations. Now, it won't proceed with the refactoring (not supported yet).
@@ -1269,7 +1271,8 @@ function doSomething(someData) {
 
 </details>
 
-[unreleased]: https://github.com/nicoespeon/abracadabra/compare/5.1.0...HEAD
+[unreleased]: https://github.com/nicoespeon/abracadabra/compare/5.2.0...HEAD
+[5.2.0]: https://github.com/nicoespeon/abracadabra/compare/5.1.0...5.2.0
 [5.1.0]: https://github.com/nicoespeon/abracadabra/compare/5.0.1...5.1.0
 [5.0.1]: https://github.com/nicoespeon/abracadabra/compare/5.0.0...5.0.1
 [5.0.0]: https://github.com/nicoespeon/abracadabra/compare/4.13.1...5.0.0